+++ changelog.md
@@ -28,6 +28,6 @@
 - Expose `objects_in_ray` through `WorldIterator::find_objects_in_ray`
 
 ## TBD
-- `World::is_body_passable` has been removed. Use the `PhysicalBody::passable` (returned by `World::body`) instead.
+- `World::is_body_passable` has been removed. Use `PhysicalBody::passable` (returned by `World::body`) instead.
 - The `passable` field on objects is respected by `NphysicsWorld`:
   Passable objects can overlap with other objects including other passable objects.